fix: Search results not being added
The search result was not being added during a users search due to a type mismatch between the value and the shecma's expectation. A fix to ensure that the schema was supported was made during this commit.
Some additional bugs that were identified while working on this also have been fixed.
Changes to error messages and their handling also have been performed